requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
bcc==0.18.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
evdev==1.4.0
executing==2.2.0
fastjsonschema==2.21.1
grimp==3.9
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
import-linter==2.3
importlib-metadata==4.6.4
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==8.10.0
nbclient==0.10.2
nbformat==5.10.4
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.8
pop-transition==1.1.2
prompt_toolkit==3.0.51
protobuf==3.12.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.0.0
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
rich==14.0.0
rpds-py==0.25.1
scapy==2.6.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.7
stack-data==0.6.3
system76driver==20.4.93
systemd-python==234
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
transitions==0.9.2
typing_extensions==4.14.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
